+++ NEWS.md
@@ -1,11 +1,22 @@
 
-# sircovid 0.10.6
+# sircovid 0.10.8
+
 * New functions to estimate effective Rt using EpiEstim
 
+# sircovid 0.10.7
+
+* Added utils for nested pmcmc
+
+# sircovid 0.10.5
+
+* Can control the number of initial infected individuals used in seeding
+
 # sircovid 0.10.3
+
 * Predefined vaccination schedule implemented
 
 # sircovid 0.10.1
+
 * Rt corrected for restructured model
 
 # sircovid 0.10.0